// src/interfaces/user.ts
// Generated by https://quicktype.io

import { FriendRequest, FriendRequestStatus } from '.';

export interface User {
    _id: string;
    name: string;
    password: string;
    email: string;
    bio: string;
    friendRequest: {
        exists: boolean;
        id: string | null;
        friendRequestStatus: FriendRequestStatus;
    };
    conversations: any[];
    friendRequestsSent: any[];
    friendRequestsReceived: any[];
    friends: { friend: User; friendRequest: FriendRequest }[];
    canReceiveFriendRequests: boolean;
    username: string;
    createdAt: string;
    updatedAt: string;
}